// src/app/shared/models/storeCustomer.ts
export interface storeCustomer {
    name:             string;
    paternal_surname: string;
    maternal_surname: string;
    type_document:    string;
    n_document:       number;
    date_of_birth:    Date;
    email:            string;
    country:          string;
    region:           string;
    province:         string;
    city:             string;
    street:           string;
    prefix:           string;
    phone:            string;
}